refactor: TitleBar 스토리 하나로 합침
